Fix: Moved Ui to app, reformat files
