+++ app/src/main/java/seng/monsters/model/ui/gui/BattleScreen.java
@@ -5,7 +5,7 @@
 //  Created by d-exclaimation on 10:09 AM.
 //  Copyright © 2022 d-exclaimation. All rights reserved.
 //
-package seng.monsters.ui.gui;
+package seng.monsters.model.ui.gui;
 
 import seng.monsters.model.BattleManager;
 import seng.monsters.model.Environment;
